+++ src/test/java/race/RaceTest.java
@@ -1,5 +1,6 @@
 package race;
 
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -12,7 +13,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 public class RaceTest {
+
     Racing racing = new Racing();
+    RacingModel racingModel = new RacingModel();
+
+    @BeforeEach
+    void setting() {
+        racingModel.prepareGame(3, 5);
+    }
 
 
     @Test
@@ -27,25 +35,8 @@ public class RaceTest {
     @DisplayName("game end test")
     @CsvSource(value = {"3:false", "5:true"}, delimiter = ':')
     void gameEndTest(int input, boolean expected) {
-       RacingModel racingModel = new RacingModel();
-       racingModel.prepareGame(3, 5);
-
        assertThat(racing.game(racingModel, input)).isEqualTo(expected);
     }
 
-    @Test
-    @DisplayName("car distance check Test")
-    void carDistanceCheckTest() {
-        CarModel car = new CarModel();
-        car.go();
-        assertThat(car.totalDistance()).isEqualTo(1);
-    }
 
-    @ParameterizedTest
-    @CsvSource(value = {"3:false", "5:true"}, delimiter = ':')
-    @DisplayName("car move Test")
-    void carMoveTest(int randomValue, boolean expected) {
-        CarModel car = new CarModel();
-        assertThat(racing.carMove(car, randomValue)).isEqualTo(expected);
-    }
 }
